# backend/src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Suporte - Sistema de Tickets",
        "version": "1.0.0",
        "description": "Documentação completa para testes de Autenticação, Gestão de Tickets e Chat de Mensagens."
    },
    "servers": [
        {
            "url": "http://localhost:3001",
            "description": "Servidor Local"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/messages/{ticketId}": {
            "get": {
                "tags": ["Mensagens"],
                "summary": "Lista o histórico de mensagens de um ticket",
                "parameters": [
                    { "name": "ticketId", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "responses": {
                    "200": { "description": "Histórico de mensagens retornado" }
                }
            },
            "post": {
                "tags": ["Mensagens"],
                "summary": "Envia uma nova mensagem no ticket",
                "parameters": [
                    { "name": "ticketId", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "content": { "type": "string", "example": "Olá, estou a analisar o seu problema." }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "Mensagem enviada" }
                }
            }
        }
    }
}